# shared-api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: 'Melanie Data API '
  description: 'A high performance & centrally cached API service for premium bots. '
  version: ' commit e2ffc88 ok @ Sep 10 11:37 pm'
servers:
- url: https://dev.melaniebot.net
  description: montreal
paths:
...
  /api/speech/stt:
    get:
      tags:
      - speech
      summary: Perform Speech To Text
      description: Return the text content of an audio file
      operationId: performSpeechtoText
      parameters:
      - description: URL of the audio or video file.
        required: true
        schema:
          type: string
          maxLength: 65536
          minLength: 1
          format: uri
          title: Url
          description: URL of the audio or video file.
        name: url
        in: query
      - description: Perform the operation with smart translation
        required: false
        schema:
          type: boolean
          title: Translate
          description: Perform the operation with smart translation
          default: false
        name: translate
        in: query
      - required: false
        schema:
          type: string
          title: Source Langs
          default: en-US,es-MX,zh-cn,uk-UA
        name: source_langs
        in: query
      - required: false
        schema:
          type: string
          title: Target Lang
          default: en
        name: target_lang
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/STTResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/speech/tts:
    post:
      tags:
      - speech
      summary: Perform Text To Speech
      description: Generate an MP3 of text input
      operationId: performTexttoSpeech
      parameters:
      - required: false
        schema:
          type: string
          title: User Id
        name: user_id
        in: query
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TTSTranslationRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TTSResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    STTResult:
      properties:
        status:
          type: string
          title: Status
        display_text:
          type: string
          title: Display Text
          default: ''
        offset:
          type: string
          title: Offset
        duration:
          type: string
          title: Duration
        detected_languages:
          items:
            type: string
          type: array
          title: Detected Languages
          default: []
        translated_text:
          type: string
          title: Translated Text
          default: ''
      type: object
      title: STTResult
...
    Speakers:
      type: string
      enum:
      - en-US-Ana
      - en-US-Amber
      - en-US-Aria
      - en-US-Ashley
      - en-US-Brandon
      - en-US-Christopher
      - en-US-Cora
      - en-US-Davis
      - en-US-Elizabeth
      - en-US-Eric
      - en-US-Guy
      - en-US-Jacob
      - en-US-Jane
      - en-US-Jason
      - en-US-Jenny
      - en-US-Michelle
      - en-US-Monica
      - en-US-Nancy
      - en-US-Sara
      - en-US-Tony
      - en-GB-Maisie
      - en-GB-Abbi
      - en-IN-Prabhat
      - en-AU-Tim
      title: Speakers
      description: An enumeration.
...
    TTSResult:
      properties:
        url:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Url
      type: object
      required:
      - url
      title: TTSResult
    TTSTranslationRequest:
      properties:
        text:
          type: string
          maxLength: 6000
          minLength: 2
          title: Text
        speaker_name:
          $ref: '#/components/schemas/Speakers'
      type: object
      required:
      - text
      - speaker_name
      title: TTSTranslationRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...